refacctor to match factory names to kernel names
